server/requirements.txt
Flask
torch
Pillow
transformers
tqdm
python-dotenv
opencv-python
geopy
firebase-admin
requests
reportlab
flask-cors